+++ R/add_race_sex.R
@@ -4,20 +4,19 @@ NULL
 
 #' Add columns for race/ethnicity and sex
 #'
-#' Add variables \code{race} and \code{sex} from \code{dbase} to a data
-#' frame, using \code{id} is the join-by variable.
+#' Add variables \code{race} and \code{sex} from \code{mdata} to a data
+#' frame, using \code{mcid} as the join-by variable.
 #'
 #' Existing variables with the same names, if any, are overwritten.
 #'
-#' @param dframe data frame
+#' @param dframe data frame with required variable \code{mcid}
 #' @param ... not used, forces later arguments to be used by name
-#' @param dbase data frame of student attributes, default midfieldstudents
-#' @return Data frame with the following properties:
+#' @param mdata MIDFIELD student data, default \code{midfielddata::student},
+#'        with required variables \code{mcid}, \code{race}, and \code{sex}
+#' @return A \code{data.table}  with the following properties:
 #' \itemize{
 #'     \item Rows are not modified
 #'     \item Columns \code{race} and \code{sex} are added
-#'     \item Data frame attributes \code{tbl} or \code{data.table}
-#'           are preserved
 #'     \item Grouping structures are not preserved
 #' }
 #' @export
@@ -25,55 +24,63 @@ NULL
 #' # TBD
 add_race_sex <- function(dframe,
                          ...,
-                         dbase = NULL) {
+                         mdata = NULL) {
+
+    wrapr::stop_if_dot_args(
+        substitute(list(...)), "Arguments after ... must be named,"
+    )
 
     # default arguments if NULL
-    dbase <- dbase %||% midfielddata::midfieldstudents
+    mdata <- mdata %||% midfielddata::student
 
-    # bind names due to nonstandard evaluation notes in R CMD check
+    # bind names due to NSE notes in R CMD check
     # var <- NULL
 
     # check arguments
     assert_explicit(dframe)
     assert_class(dframe, "data.frame")
-    assert_class(dbase, "data.frame")
+    assert_class(mdata, "data.frame")
+
+    # The dframe argument is modified "by reference." Thus changing its value
+    # inside the function immediately changes its value in the calling frame
+    # --- a data.table feature designed for fast data manipulation,
+    # especially for data that occupies a lot of memory.
+    setDT(dframe)
+    setDT(mdata)
 
     # existence of required columns
-    assert_required_column(dframe, "id")
-    assert_required_column(dbase, "id")
-    assert_required_column(dbase, "race")
-    assert_required_column(dbase, "sex")
+    assert_required_column(dframe, "mcid")
+    assert_required_column(mdata, "mcid")
+    assert_required_column(mdata, "race")
+    assert_required_column(mdata, "sex")
 
     # class of required columns
-    assert_class(dframe[, id], "character")
-    assert_class(dbase[, id], "character")
-    assert_class(dbase[, race], "character")
-    assert_class(dbase[, sex], "character")
+    assert_class(dframe[, mcid], "character")
+    assert_class(mdata[, mcid], "character")
+    assert_class(mdata[, race], "character")
+    assert_class(mdata[, sex], "character")
 
-    # preserve original data.frame, data.table, or tibble class
-    df_class <- get_dframe_class(dframe)
-
-    # prepare dframe, preserve column order for return
-    # omit existing column(s) that match column(s) we add
-    setDT(dframe)
-    added_cols <- c("race", "sex")
+    # preserve column order except columns that match new columns
     names_dframe <- colnames(dframe)
+    added_cols <- c("race", "sex")
     key_names <- names_dframe[!names_dframe %chin% added_cols]
     dframe <- dframe[, key_names, with = FALSE]
 
-    # work on the dbase data
-    cols_we_want <- c("id", added_cols)
-    rows_we_want <- dbase$id %chin% dframe$id
-    DT <- dbase[rows_we_want, cols_we_want, with = FALSE]
+    # work on the mdata data
+    cols_we_want <- c("mcid", added_cols)
+    rows_we_want <- mdata$mcid %chin% dframe$mcid
+    DT <- mdata[rows_we_want, cols_we_want, with = FALSE]
 
     # left-outer join, keep all rows of dframe
-    dframe <-  merge(dframe, DT, by = "id", all.x = TRUE)
+    dframe <-  merge(dframe, DT, by = "mcid", all.x = TRUE)
 
-    # prepare return
-    # order columns and rows using original columns as keys
+    # restore column and row order
     set_colrow_order(dframe, key_names)
 
-    # restore original data frame class
-    revive_class(dframe, df_class)
+    # remove grouping structure, if any
+    setkey(dframe, NULL)
+
+    # enable printing (see data.table FAQ 2.23)
+    dframe[]
 }